fix bug
Former-commit-id: 87bd97a067bd356d06e2332633734a9140e6b1ac [formerly 7048181b7cd45fa3893b02804dfecdf756052ec9] [formerly 96cadc33bf0384255afc6e4a8bd00c4c64cb8b63]
Former-commit-id: e12d00baa1896811dca082a7afa1c7d0b5ba4b5b
Former-commit-id: 781b35b0a212edd7c6afba28cfbaab3da262ddcf

+++ gmql/dataset/DataStructures/RegField.py
@@ -140,7 +140,7 @@ class RegField:
             other = self.exp_build.getREType("string", other)
         elif isinstance(other, int):
             other_name = str(other)
-            other = self.exp_build.getREType("int", str(other))
+            other = self.exp_build.getREType("float", str(other))
         elif isinstance(other, float):
             other_name = str(other)
             other = self.exp_build.getREType("float", str(other))
